new file:   client/src/components/Dash.jsx 	modified:   client/src/pages/Dashboard.jsx
